DDF example using files as devices, initial commit
This is an exercise to learn how to implement drivers using DDF.
The devices are files in one directory, and the idea is that a driver
would RX while reading from a file, and TX while writing to it.
Drivers would implement reading and writing to pcap files or other
formats.

Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>