physmem: change API, alloc now does a mapping too
We add the _reserve() function to the API to do what was previously
achieved with _alloc(), which is to reserve a set of huge pages that are
physically contiguous.
We changed _alloc() so it now calles _reserve() and _map() so it returns
a block that is already mapped in the process' virtual address space.

Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>